// src/models/models.ts
export interface User {
    id?: number;
    name?: string;
    username?: string;
    email?: string;
    address?: Address;
    phone?: string;
    website?: string;
    company?: Company;
    posts?: Post[];
}

export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: Geo;
}

export interface Geo {
    lat: string;
    lng: string;
}

export interface Post {
    id?: number;
    userId?: number;
    title?: string;
    body?: string;
}

export interface Company {
    name?: string;
    bs?: string;
    catchPhrase?: string;
}